Major (but backwards compatible) update to the status tests to make them more efficient.
Previously, every status test was executed regardless of its
necessity in determining the status of the nonlinear iteration. This
is the inefficiency that this change addresses.

There is now an additional checkStatus function that takes a second
argument of an enum type called CheckType. This indicates what type of
check should be performed by checkStatus. The choices are "Complete",
"Minimal", and "None".

Complete - Every status test and corresponding subtest should be
performed. You might call checkStatus with this argument on the final
solution.

Minimal - Only the minimal number of status tests necessary to
determine the outcome. This is the default.

None - No evaluation should be done, unless it is necessary that the
test be done at every iteration. One example is the Stagnation test -
it has to be executed at every iteration.

If you have any of your own status tests, you may wish to update them
to use the two-argument version of checkStatus.