Major changes include the following:
1) Addition of 3 new methods (Feng-Pulliam method and the "2" and "2+"
   methods that I have developed as part of my thesis research)

2) New options related to restarting, orthogonalization, and a
   "shortcut method"

3) Fixes to preconditioning (right not available until fixes in
   ApplyRightPreconditioning)

4) Streamlining of certain sections of code.

Other changes include making the code compatible with parameter lists
and return status.  This commit also fixes bug #144 from bugzilla.

For the moment, I am not intending for these new methods to be
available for general use, but I am committing the code to have an
updated version of the "Tensor Based" solver, which uses the basic
"Tensor" direction.  See the Broyden example problem for what is
supported.

Doxygen comments have not been updated.  They will be updated sometime
once the methods have been tested and we decide how to organize this
code within NOX.