Fixes bug #285
Currently if nox is being used in a time stepping code, the reset
function is being called at the beginning of each time step.  This
forces the parameter list to be re-parsed and the group to be recloned
etc...  This can be highly inefficient and in the case of Xyce has led
to about 5% increase in wall clock time.

A light-weight reset method has been added that does not change
anything except to reset the counters in the nonlinear solver.  This
can be dangerous since the status tests, line searches, and inexact
newton method will not know about the reset.  I will fix these methods
before the bug is closed out.