----------------------------------------------------------------------
 Modified Files:
 	AztecOO.cpp
 ----------------------------------------------------------------------

Fixes bug 304.

Formerly AztecOO reset all the options and params values whenever a new
Epetra_LinearProblem was set using the SetProblem() method.  This was not
the best approach.  Instead the options and params are now set only if it
is the first time a problem has been set.  On subsequent calls to SetProblem,
the existing options and params are used.  If the PDL or Symmetry of the
problem changes on subsequent calls to SetProblem, a warning error code of 1
is generated.

Test were run and passed.