Fixed compiling nox on cplant.  We had a bunch of huge problems with our configure script:
1.  Moved the location of the check for libmpi.a.  The check only
occurs if we are compiling the examples that need libmpi.a (i.e. the
epetra examples).  The problem was that the flag for whether we are
compiling the epetra examples was being set AFTER this check.
Consequently, we were never checking for libmpi.a.  This caused nox to
fail to set -lmpi in the $(LIBS) variable.  On normal linux mpich this
doesn't matter, since the mpiCC compiler script automatically includes
-lmpi.  On cplant and asci-red this check was failing and we never got
-lmpi set in the links for examples.

2.  While nox uses CXXREPOS to include DEC template subdirectories,
the rest of Trilinos uses the flag TRILINOS_TEMPLATE_OBJS.  For
trilinos users this requires nox compiles to set both flags
redundantly.  I have added a check for TRILINOS_TEMPLATE_OBJS which
sets CXXREPOS in nox to that value.

3. Fixed some of the library makefiles to also include $(XTRALDADD) on
the link lines.  This is standard in most of the other packages.