----------------------------------------------------------------------
 Modified Files:
 	src/Epetra_SerialDenseMatrix.cpp
 	src/Epetra_SerialDenseMatrix.h test/SerialDense/cxx_main.cpp
 ----------------------------------------------------------------------

Closes bug 366.

Added a new method called Scale() that allows a user to scale a matrix
in-place using the value passed in to the Scale() method.

Also added two new methods, NormOne and NormInf that are identical to
OneNorm and InfNorm.  The new methods are compatible in name with the
same methods used in other Epetra matrix classes.

Added tests to the SerialDense test driver.  All tests pass.