Implemented the "pseudo-solver wrapper" idea for dealing with status tests in LOCA discussed in the last NOX meeting.  This involved a few changes to the LOCA abstract group hierarchy and the addition of some new files:
1)  Added a LOCA::Extended::AbstractGroup that provides a generic interface
for extended (a.k.a. super) groups.  The interface provides methods for
retrieving the immediate underlying group (getUnderlyingGroup()) and the
base-level underlying group (getBaseLevelUnderlyingGroup()) which is to return
the lowest-level underlying group in a family of nested super groups.  This
class now servers as a parent class for all extended continuation/bifurcation
groups.  Note that this is a change in the meaning of getUnderlyingGroup()
which previously did what getBaseLevelUnderlyingGroup() does now.

2) Changed all of the extended continuation/bifurcation extended groups to
use the new extended abstract group base class.

3) Added a LOCA::Solver::Wrapper class which implements the pseudo-solver
idea.  It takes a NOX solver as an argument in its constructor and forwards
all NOX::Solver::Generic methods to this solver, except the getSolutionGroup()
and getPreviousSolutionGroup() methods.  These instead return the immediate
underlying group of the current and previous solution groups (via
getUnderlyingGroup()) if the solution groups are extended (derived from the
LOCA::Extended::AbstractGroup).

4) Added a LOCA::StatusTest::Wrapper class which wraps a NOX status test.  The
checkStatus() method takes the passed solver and creates a pseudo-solver from
it using the LOCA::Solver::Wrapper class and then passes this solver to the
wrapped status test.  This allows the wrapped status test to get the
proper solution groups from the solver.

5) Created several status tests for the parameter and null vector components
of the continuation/bifurcation extended groups.  These status tests implement
the convergence criteria used by LOCA 1.0 in Salsa.

6) Modified the stepper to use the new status tests.  There are now two
stepper constructors which take one or two status tests.  If one status test
is provided, it is assumed that status test will work with or without the
continuation extended group wrapping the supplied group (such as NormF).
If two are provided, the first one is assumed to be for the solution component
of the supplied group (which could be a bifurcation group, in which case
the solution component means the entire solution, e.g., [x y mu] for a turning
point) and the second is for the parameter component of the continuation group.
A final status test is created by creating and AND combo of the two.   This
separation was necessary since the first step of a continuation run applies the
NOX solver to the supplied group to compute the first point on the group with
out using a continuation group.

An example prroblem as well as better documentation describing how to use
these status test wrappers will be coming soon.