Added 2 new files:
1. NOX_SharedObjectTemplate.H - We use shared objects in all support
groups for the Jacobian, preconditioner, and/or the entire linear
system.  I just made one object that we can all use instead of
rewriting the shared object for each support group.  This is currently
used in the new epetra support that will be checked in soon.

2. NOX_Direction_Utils_InexactNewton.* - Broke the inexact forcing
term calculations out of the Newton direction object.  It is now a
utility object.  I needed access to these methods in other parts of
NOX (Inexact Trust Region in PRERELEASE).  The Newton direction does
not call this yet but I will switch it over soon.