----------------------------------------------------------------------
Committing in .

Modified Files:
	Makefile.am Makefile.in aclocal.m4 configure
	example/Makefile.in src/Makefile.in
----------------------------------------------------------------------
Added a number of m4 and autotools files to the list of files to
include in the distribution tarball.  We want to distribute a tarball version
of new_package.  In the future, we may even distribute a new_package
tarball apart from the rest of Trilinos.