----------------------------------------------------------------------
 Committing in .

 Modified Files:
 	aclocal.m4 configure configure.ac
----------------------------------------------------------------------
Bug 421.

The AC_FUNC_MALLOC macro only recognizes malloc functions that are compatible
with the GNU C library.  The malloc function on DEC machines is not compatible
with the GNU C library.

We will now check for malloc, if that fails, we will run AC_FUNC_MALLOC.
Autoconf will find malloc on the DEC machines.

This fix worked for AztecOO.