this is a config file that has the definitions to build zoltan with a a CCA component version of it too.
The reason that a new config file had to be made - rather than keeping the
old Config.linux file and addind to it are :

a) The original config file has NEMESIS uncommented i .e. on linux one
builds NEMESIS based tests. This is no longer working after RedHat 8.0.
NEMESIS requires the linking of libnetcdf.a which was compiled 3 years ago.
Compiled into it is a symbol __ctype_b, which is no longer supplied by
modern libc.a. Thus libnetcdf is unusable until compiled again.
I have thus disabled the compilation and use of NEMESIS in zdrive but
commenting out NEMSIS_* in this new config file

b) By default, on the Sandia network, on linux machines, zoltan uses
LAM mpi that comes default with linux. I have yet to compile the CCA
framework with LAM - it has been compiled with MPICH. Hence I've had to change
mpicc in Config.linux

c) I have had to add to other keys in Config.linuc_cca - CCA_SPEC_DIR and
DCCAFE_DIR to indicate where the CCA spec and framework are.


Once libnetcdf is recompiled and I compile the CCA framework with LAM mpi,
we can merge c) into the usual Config.linux file and this file can disappear.
Till then we keep this and compile the zoltan component as

gmake ZOLTAN_ARCH=linux_cca zoltan_component. This will remake zoltan, zdrive
and the zoltan_component.


THE ZOLTAN COMPONENT IS NOW READY FOR USE. GO AHEAD AND CHECK IT OUT.