Fixes to configure so loca will compile on some platforms.
When libloca.a became dependent on lapack, the Fortran wrappers for
lapack automatically drew in a configure check on c2f to make sure the
wrappers would compile under c.  This check was failing certain builds
due to missing c compiler flags.  Most Trilinos packages check the
--with-cflags configure command, but NOX was not.  So some users were
using this flag for configuring all of trilinos, but the flag was
ignored by nox.  I have added checks for both --with-cflags and
--with-fflags.