----------------------------------------------------------------------
Committing in .

Modified Files:
	test-harness
----------------------------------------------------------------------
A mod to provide a more descriptive subject line than "At least one test
failed".  Now we should be able to at least tell if a test failed, a
compile error occurred, or an update conflict occurred.  I am going to work
on supplying the scirpt name or something more descriptive when a test fails.
Currently, if I do so, the subject line can be hundreds of chars long.