Added a few lines to Epetra_BlockMap::operator= so that the data member BlockMapData_->Directory_ is re-generated. Previously, Directory_ was left holding the map that had just been replaced. This was causing my column-permuted objects to crash...
Fixed a couple of minor memory leaks in MapColoring/cxx_main.cpp.

Removed a couple of unnecessary dynamic_casts in Epetra_MpiComm.cpp.
You don't need dynamic_cast to assign a derived class instance to a
base class pointer. And in fact, using dynamic_cast in this way causes
some debuggers to be unable to examine the resulting base-class instance.